Revise post page to add post header
